baseclass for request and response
