+++ Titanium.Web.Proxy/Decompression/DecompressionFactory.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Titanium.Web.Proxy.Http;
 
 namespace Titanium.Web.Proxy.Decompression
 {
@@ -18,9 +19,9 @@ namespace Titanium.Web.Proxy.Decompression
         {
             switch (type)
             {
-                case "gzip":
+                case KnownHeaders.ContentEncodingGzip:
                     return gzip.Value;
-                case "deflate":
+                case KnownHeaders.ContentEncodingDeflate:
                     return deflate.Value;
                 default:
                     return def.Value;
