+++ Titanium.Web.Proxy/Http/KnownHeaders.cs
@@ -40,6 +40,8 @@ namespace Titanium.Web.Proxy.Http
 
         // Response headers
         public const string ContentEncoding = "content-encoding";
+        public const string ContentEncodingDeflate = "deflate";
+        public const string ContentEncodingGzip = "gzip";
 
         public const string Location = "Location";
